Changed cost to float

+++ expense-tracker/src/main/java/com/example/expense_tracker/domain/dto/ExpenseRequest.java
@@ -10,7 +10,7 @@ public record ExpenseRequest(
    String note,
    @NotNull
    @Positive
-   Integer cost,
+   Float cost,
    @NotNull
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
    LocalDate expenseDate,
